discloud/import/1271746295496704093/requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
requests